pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toteat-integration"
version = "0.1.0"
description = "Toteat to PostgreSQL ingestion pipeline"
requires-python = ">=3.11"
dependencies = [
  "psycopg[binary]>=3.2",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
toteat-sync = "toteat_integration.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
